Support multi-channel interfaces
Prior to this commit, if a device with more than eight channels appeared
in a device list, only the first eight channels were actually displayed.
Devices below the multi-channel device were not displayed.

The issue relates to the naming of channels; they were given names
depending on their channel number, which works well for up to eight
channels. When the 9th channel appeared, it was considered to be
Bar.NONE, which caused its name to be displayed in the center of the
screen. Then, retreiving the name caused an exception, which caused the
remaining devices not to be displayed.

The solution is to give Bar.NONE a value that can't possibly be confused
for a channel index, and to default to displaying the channel number if
there is no name for the channel.