Always display non-mono channel numbers
Prior to this commit, channel numbers were only displayed if there was
no name for the channel, defined in Screen.SIDES. In reality, channels
numbers don't necessarily correspond to a speaker function, e.g. in a
music studio context. Then it's more useful to know the actual channel
number.

With this commit, channel numbers are always displayed, with the speaker
side name displayed if available.